move fiat reverse engineering into fiatinterface
